Automatic Speech Recognition day3/requirements.txt
openai-whisper
faster-whisper
torch
ffmpeg-python
pyaudio
vosk
tensorflow-cpu==2.12.0
scikit-learn==1.3.0
numpy==1.23.5 
pandas==2.0.3